+++ addin/APBridgeAddIn/IpcModels.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
+#nullable enable
+
 namespace APBridgeAddIn
 {
     public record IpcRequest(