+++ addin/APBridgeAddIn/Module1.cs
@@ -1,7 +1,5 @@
 using ArcGIS.Desktop.Framework;
 using ArcGIS.Desktop.Framework.Contracts;
-using System;
-using System.Threading.Tasks;
 
 namespace APBridgeAddIn
 {
@@ -17,14 +15,17 @@ namespace APBridgeAddIn
         {
             _bridgeService = new ProBridgeService("ArcGisProBridgePipe");
             _bridgeService.Start();
-            System.Diagnostics.Debug.WriteLine("[APBridgeAddIn] Named pipe server started on ArcGisProBridgePipe");
             return true;
         }
 
         protected override bool CanUnload()
         {
+            return false;
+        }
+
+        protected override void Uninitialize()
+        {
             _bridgeService?.Dispose();
-            return true;
         }
 
         public void RestartBridge()